Fixed unreliable enabling of doc snippets
Testing for QT.widgets.name in qtmultimedia.pro causes the behavior of
that .pro file to depend on whether or not qtbase.pro has been
processed.  Due to QTBUG-23376, and because these are sibling .pro
files, this is not safe - nothing guarantees that qtmultimedia.pro is
processed after qtbase.pro.

Work around the issue by pushing the QT.widgets.name test down one
level.

Change-Id: I3e155b3f18ad5fa8c527cf5a806313db4b6aabeb
Reviewed-by: Jonas Rabbe <jonas.rabbe@nokia.com>