Fix the QML list interface
The QML list interface was changed, so that there is now a read only
constructor and the default properties are removed. This is because the
original interface was very runtime centric, like other parts in QML.
The new interface was changed for the needs of the QML designer and
debugger.

Task-number: QTBUG-9390

Change-Id: Ia61a2457432b4fc5c0cf440eeebd9d3a6c4c6463
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>