QAudioInput/Output documentation: Fix slot naming in snippets
Having a code snippet use a slot (stateChanged) with a name
identical to a signal in QAudioInput / QAudioOutput classes causes
problems (incorrect links) in class documentation that refer to
the snippet code.

This change renames the slots according to standard Qt convention.
Same change is applied to example applications as well.

Task-number: QTBUG-26688
Change-Id: I0c6181240237d0c642b73fe18f3ddb5137b7f207
Reviewed-by: Dmytro Poplavskiy <dmytro.poplavskiy@gmail.com>
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>