Fix file path bug in audiorecorder example
Default path and name of the output wave audio files were wrong.  This
means the files did not have the right name to be opened on Windows, and
did not have the right path because they were being saved to the root
folder. (C:/ on windows, and / on OSX).  Now we fall back to the default
filepath and name when we try to record without giving a path and
filename.

Task-number: QTBUG-28663
Change-Id: I8c782a31710cac65cd3c1ecd44de10b41c46e93d
Reviewed-by: Christian Stromme <christian.stromme@digia.com>