Android: Fix compilation on Windows
The qtCompileTest function in configure.prf does not work when
cross-compiling on Windows because it detects the wrong make executable.
And it's anyway not the best way to do this, since what we really want
to check for is the existence of the SDK version, whereas the compile
test could potentially hide a lot of other problems under the same
error message.

Task-number: QTBUG-31405
Change-Id: I2c16b214e40bac5751d4e7098edcc30318756621
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>
Reviewed-by: Christian Stromme <christian.stromme@digia.com>