Tell glib not to complain about the use of API recently deprecated
When building with glib 2.32 or later, without this #define, it prints a
warning saying:

function "g_value_get_char" was declared deprecated ("Use 'g_value_get_schar' instead")

Since g_value_get_schar is new in 2.32, we can't use it without #ifdef
around the call point. Since the old and new functions are identical in
behavior, just ask glib not to complain about the use of old functions.

I don't know which is the minimum version of glib we require, so I went
for the lowest available macro.

Change-Id: I33f9aa8497fc6bd50fffb3c59ee7e8e7fcdfe110
Reviewed-by: Yoann Lopes <yoann.lopes@digia.com>