Improvide device selection in the video filter example
The non-OS X path also uses clGetGLContextInfoKHR when available,
instead of just blindly taking the first GPU device for the platform.

This way we get the correct OpenCL device in case the application is using an
OpenGL context on the same vendor's second GPU.

On Windows desktop OpenGL is now correctly forced and therefore there is
no need to waste time on runtime checks for ANGLE later on.

Remove the hard-coded custom pathes from the .pro file.

Also fixes error reporting on the OS X specific path.

Change-Id: I8d6ab867510d113d5135b61e66822381e81c995a
Reviewed-by: Yoann Lopes <yoann.lopes@theqtcompany.com>